+++ src/features/optional/major/fastClear/callbacks/postNPCUpdate.ts
@@ -20,7 +20,10 @@ export function dingle(npc: EntityNPC): void {
 
   // Fix the bug where a Dangle spawned from a Brownie will be faded. We only care about Dangles
   // that are freshly spawned.
-  if (npc.Variant === DingleVariant.DANGLE && npc.State === NpcState.INIT) {
+  if (
+    npc.Variant === (DingleVariant.DANGLE as int) &&
+    npc.State === NpcState.INIT
+  ) {
     npc.SetColor(Color.Default, 1000, 0, true, true);
   }
 }
