+++ src/features/optional/quality/combinedDualityDoors.ts
@@ -72,7 +72,7 @@ export function postPEffectUpdate(player: EntityPlayer): void {
 }
 
 function getPlayerOnDevilSide(player: EntityPlayer, door: GridEntityDoor) {
-  const useYAxis = door.Slot % 2 === 0;
+  const useYAxis = (door.Slot as int) % 2 === 0;
   const invertDirection = shouldInvertDirection(door.Slot);
 
   // We combine position and velocity to project where the player will be a frame from now. We do
